fix amount

+++ src/main/java/com/iker/Lexly/Entity/Order.java
@@ -23,7 +23,7 @@ public class Order {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
